requirements.txt
streamlit==0.89.0
Pillow==8.3.2
numpy==1.21.2
tensorflow==2.6.0


To generate the requirements.txt file, you can follow the steps I mentioned earlier:

1) Open your terminal or command prompt.
2) Navigate to the directory of your Streamlit app.
3) Activate your virtual environment if you're using one.
4) Run the following command to generate the requirements.txt file:

- pip freeze > requirements.txt

